Redid the marker system
Each of the functions that transform/edit elements now put all their
changes into a single container element, on which they store a piece of
information that can be used to detect whether it needs to be
re-generated. If this information (the key) changes the container
element is completely destroyed and re-created.