Improved layout switching
Instead of a toggle area at the bottom of the screen there now is a menu
to switch between the different layouts. This also allows you to select
from any of the 3 available layouts, instead of toggling between default
and one other depending on the screen size.

The layout is now also kept between page loads.