Fix layout positioning on different screen sizes
There were two problems causing the layouts to be off depending on the
resolution.

The first is that the width of the boxes depends on the resolution, so
on wider resolutions the positions of the boxes would be off in some
layouts.

The second is that the height of the header also depends on the
resolution, so on wider resolutions there would not be enough space at
the bottom for layouts that use the full height. In addition the margin
left at the bottom for such layouts was inconsistent.

This change fixes all of these issues, making all layouts now work
properly on a larger variety of screen sizes.