Integrated with built-in container management
Now that basic container management is built into D&D Beyond there is
not need for this extension to track this, so it doesn't. The built-in
offering is still lacking in some departments though, such as ignoring
the weight of items in some containers (like a Bag of Holding), so this
functionality remains in place.