Remove inventory/equipment management
Rejoice! This is now (finally) natively supported by D&D Beyond.

Well, mostly. It doesn't support custom items in containers (yet), and
it doesn't display a few pieces of (mostly irrelevant) information (like
the total cost of all items in a container), but it's good enough that
this bit of functionality can be retired from this userscript.