Redid sticky headers
The contents now no longer extend behind the sticky headers, which looks
_much_ better without looking weird with transparant backgrounds.

Still a work in progress, only works for the desktop view for now.

The downside is that it is no longer pure CSS. There is now a setting to
disable it since it's a bit more involved (and thus more prone to
breaking) than before.